[input] Expand ContenderId and StreamId to 64-bit
scenic_impl::input::StreamId and scenic_impl::input::ContenderId are
migrated to 64-bit integers to support larger identifier spaces and
prevent wrapping.

Bug: 519648265, 519649047
Test: fx test input_unittests
TAG: agy
CONV: 40b2de25-4db2-4a73-808a-adba9c8842e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1696596
Fuchsia-Auto-Submit: Caroline Liu <carolineliu@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>

+++ src/ui/scenic/lib/input/gesture_contender.h
@@ -27,7 +27,7 @@ enum GestureResponse {
   kUndefined = 9
 };
 
-using ContenderId = uint32_t;
+using ContenderId = uint64_t;
 constexpr ContenderId kInvalidContenderId = 0;
 
 // Interface for a gesture disambiguation contender. All methods are called in response to
